fix: moved cache view
