+++ commerce_coordinator/apps/stripe/views.py
@@ -6,14 +6,12 @@ import logging
 import stripe
 from django.conf import settings
 from django.views.decorators.csrf import csrf_exempt
-from edx_django_utils.cache import TieredCache
 from rest_framework import status
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
-from rest_framework.views import APIView
 
 from commerce_coordinator.apps.core.constants import PaymentState
-from commerce_coordinator.apps.core.memcache import safe_key
+from commerce_coordinator.apps.core.views import SingleInvocationAPIView
 from commerce_coordinator.apps.rollout.utils import is_commercetools_stripe_refund, is_legacy_order
 from commerce_coordinator.apps.stripe.constants import StripeEventType
 from commerce_coordinator.apps.stripe.exceptions import (
@@ -28,72 +26,6 @@ logger = logging.getLogger(__name__)
 stripe.api_key = settings.PAYMENT_PROCESSOR_CONFIG['edx']['stripe']['secret_key']
 endpoint_secret = settings.PAYMENT_PROCESSOR_CONFIG['edx']['stripe']['webhook_endpoint_secret']
 source_system_identifier = settings.PAYMENT_PROCESSOR_CONFIG['edx']['stripe']['source_system_identifier']
-NOTIFICATION_CACHE_TTL_SECS = 60 * 10  # 10 Mins
-
-
-class SingleInvocationAPIView(APIView):
-    """APIView that can mark itself as running or not running within TieredCache"""
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.meta_id = None
-        self.meta_view = None
-        self.meta_should_mark_not_running = True
-
-    @staticmethod
-    def _view_cache_key(view: str, identifier: str) -> str:
-        """Get cache key for view and identifier"""
-        return safe_key(key=f"{view}_{identifier}", key_prefix="ct_sub_msg_invo", version="1")
-
-    def mark_running(self, view: str, identifier: str, tf=True):
-        """Mark view as running or not running"""
-        self.set_view(view)
-        self.set_identifier(identifier)
-        key = SingleInvocationAPIView._view_cache_key(view, identifier)
-
-        if TieredCache.get_cached_response(key).is_found or not tf:
-            try:
-                TieredCache.delete_all_tiers(key)
-
-            # not all caches throw this but a few do.
-            except ValueError as _:  # pragma no cover
-                # No-Op, Key not found.
-                pass
-
-        if tf:
-            TieredCache.set_all_tiers(key, tf, NOTIFICATION_CACHE_TTL_SECS)
-
-    @staticmethod
-    def _is_running(view: str, identifier: str):
-        """Check if view is running"""
-        key = SingleInvocationAPIView._view_cache_key(view, identifier)
-        cache_value = TieredCache.get_cached_response(key)
-        if cache_value.is_found or cache_value.get_value_or_default(False):
-            logger.debug(f'[CT-{view}] Currently processing request for %s, ignoring invocation', identifier)
-            return True
-        return False
-
-    def set_view(self, view: str):
-        """Set the view to mark as running"""
-        self.meta_view = view
-
-    def set_identifier(self, identifier: str):
-        """Set the identifier to mark as running"""
-        self.meta_id = identifier
-
-    # Right now we DON'T want to mark the view as not running, unless error.
-    # def finalize_response(self, request, response, *args, **kwargs):
-    #     tag = self.meta_view
-    #     identifier = self.meta_id
-    #     if self.meta_should_mark_not_running:
-    #         SingleInvocationAPIView.mark_running(tag, identifier, False)
-    #     return super().finalize_response(request, response, *args, **kwargs)
-
-    def handle_exception(self, exc):
-        """Mark view as not running on exception"""
-        tag = self.meta_view
-        identifier = self.meta_id
-        self.mark_running(tag, identifier, False)
-        return super().handle_exception(exc)
 
 
 class WebhookView(SingleInvocationAPIView):
